+++ src/Http/Graphql/typeDefs/Input/TradingEngineSymbol__Input.js
@@ -34,7 +34,7 @@ module.exports = gql`
     askSpread: Float!
     stopsLevel: Int!
     lotSize: Int!
-    percentage: Int!
+    percentage: Float!
     baseCurrency: String
     quoteCurrency: String!
     symbolType: TradingEngine__SymbolTypes__Enum!
@@ -82,7 +82,7 @@ module.exports = gql`
     askSpread: Float!
     stopsLevel: Int!
     lotSize: Int!
-    percentage: Int!
+    percentage: Float!
     baseCurrency: String
     quoteCurrency: String!
     symbolType: TradingEngine__SymbolTypes__Enum!
